refactor: api - define project structure to conform DDD architecture
