+++ mocked-api/api/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 from strawberry.fastapi import GraphQLRouter
-from schema import schema
+from api.schema import schema
 
 graphql_app = GraphQLRouter(schema)
