+++ mocked-api/api/schema.py
@@ -1,7 +1,7 @@
 import strawberry
 from typing import Iterable
 from strawberry import relay
-from schema_types.blood_donation_center import BloodDonationCenter
+from domain.entities.blood_donation_center import BloodDonationCenter
 from mocked.mocked_blood_dotation_centers import load_bdc_data
