+++ mocked-api/application/exceptions.py
@@ -1,0 +1,1 @@
+# For application-specific exceptions, if any
