+++ mocked-api/application/services/fetch_blood_donation_centers_service.py
@@ -1,0 +1,1 @@
+# Handles the logic of fetching and returning data
