+++ mocked-api/domain/enums/blood_dotation_center_category.py
