+++ mocked-api/domain/exceptions.py
@@ -1,0 +1,1 @@
+# For domain-specific exceptions, if any
