+++ mocked-api/domain/value_objects/weekday.py
