+++ mocked-api/resources/output/bloodDonationCenters/bloodDonationCenters-query.graphql
